Re-implemented worldLoadingDeferred.
This should help with #1, but won't fix it entirely.  The world saves
right, but LongHashMap.Entry still isn't found.  (I'm thinking that that
is because all of the chunks are `noLongerNeeded` instead of still
getting batched up, and thus the list is empty and the entry can't be
found.  So it's not an issue as everything is already saved.  But I'm
not sure.  Also, I want the saving screen to appear...)