Handling error conditions when importing tile entities.
Instead of catching the exception directly.  The main condition that an
exception was thrown for was the DataInputStream being null (IE, the
chunk hadn't been saved before), but we can just check this rather than
having an NPE be thrown, allowing other errors to still be logged.

Which brings up a second error -- the TileEntity had a null world,
meaning that it would still crash, and the fact that the error wouldn't
be logged meant that nobody would know.  I've fixed that.

Relates to #3.  It's still not fixed yet, though.