Improved messages for chunk saving.
This also moves the ThreadedFileIOBase code to the saveEverything
section, so that it's not skipped.  (It being skipped was the reason the
GUI still hid automatically in #3 -- it was set as done working
imediately).  Also, I fixed the emptyness check, so that people won't be
confused by "Failed to get class for LongHashMap.Entry".