New onRemoveEntityFromWorld code that doesn't keep entities loaded!
Instead, it builds up a list of new entities and uses that while saving.

This entirely fixes #5, although there are some performance things I can
still do (namely, caching the chunk position).  And I'll probably want
to make an in-game GUI for controlling entity save distances (probably
put it in the same GUI as for #4), since a lower renderdistance on the
server will cause them to despawn regardless of the actual distance, and
thus not get saved.