OK, I'm pretty sure I fixed #7 and fixed #8 this time.
I redid the large chest contents.  I think what was happening before is
that the data for tile entity #2 was getting put in the slot for TE #1
(and vice versa) when one of the sides was getting clicked
(specifically, the south on an east/west facing chest), and minecraft
just didn't like that.  Plus, the issue from 12d5bd9 (positions getting
mixed up) also caried caused issues.