Fixes for things that were sometimes null before world load.
Somehow, ticks ocasionally occured before world loading.  I don't know
why.  I've added a null check before accessing the player for gui
closing (which fixes a crash I sometimes had, but not often), and set a
default value for `receivedPackets` so that that isn't null (fixing an
issue soccerguy3 had, although I'm not sure if it fixes the base cause).