Created IWDLMod, which receives WDL's events.
This should be good for liteloader, but it doesn't actually care how the
mod is loaded, only that it's added to the list in WDLApi and that it's
only added once.  Right now it receives most of the common events;
later, I might add the ability to specify entity ranges.

It also suports the profiler.

This also includes a testmod, which will be removed later.