Temporarilly allow moving while the save progress GUI is open.
It's an interesting idea... but it doesn't seem to work well.  I'll
experiment further, but this is going to be reverted.

While keyboard input works and placing blocks works, breaking them
doesn't; you can open the chat GUI which breaks _everything_; things
just seem buggy in general.  I don't like the risk of having a imobile
GUI for saving, but it seems like it's the best plan for now (just
gotta' beware of death).