Add automatic converter
To see what's translating, use Firefox (recommended for viewing), or another common browser (you'll need to download the driver from the url in the error raised).

For automatic without viewing, download PhantomJS and use that instead.

You need to set up the language mapping to see what to put at the end of the google translate url

To use:

    py -3
    import langconvert
    lanconvert.translate('en_US', yournewlang)

:D