Fix major bug where player info was not properly saved.
I'm not sure how I didn't catch this.  But basically, in the past, the
player NBT was written into level.dat, in adition to being saved in the
playerdata folder.  Then (probably when refactoring the player saving
code in 9e008d518dcc257240ebb55aae9ec18ce553a265), I moved it into a
separate method, and passed it an empty NBTCompound, thinking that the
parameter was actually just a cloned tag.  I've fixed it now.