Don't fix entity positions if their server position is 0,0,0.
Fixes uyjulian/LiteModWDL#4.  Also fixes the issue were the world would
sometimes randomly go invisible (which is why the world was refreshed
after saving finishes).

Basically, due to boats being awesome (http://imgur.com/3QQchZL.gifv)
(the same thing used to happen to squid and other mobs (especially when
packed tightly)), I added a system where entities would be returned to
their server positions when saved.  However, I didn't factor in the fact
that some entities had special spawn packets (mainly, paintings -
http://wiki.vg/index.php?title=Protocol&oldid=7121#Spawn_Painting).
This means that the serverPosX, serverPosY, and serverPosZ were all 0 -
the entity was teleported to 0,0,0.  The same thing also happened to
players, but for players, they were teleported back 1 frame (or less)
later.  But this 1 tick away causes some chunks to visually unload, and
breaks a lot of things.

What is done now is that it only fixes the position if the server
positions are set - in other words, one of them must not be 0.  (There
is a small chance an entity is actually spawned at 0,0,0, but I consider
that unlikely enough that it's not worth trying to fix their position in
that case).  I also explicitly skip players.