Stop reloading chunks after the stopping download
This was used to fix the world vanishing, but since the root cause of
that seems to have been solved, it should no longer be necessary.  This
reload doesn't cause too much of a performance issue on my computer, but
for others it may generate a largish lag spike.  (FYI, reloading chunks
is done with F3+A normally).