Don't save empty chunks
(Broken) servers may send an unload chunk packet twice, or else mess up
on chunk unloading.  If we try to save that chunk, empty data (or worse)
gets saved.  Chunks are now checked as non-empty (and in the correct
location) to avoid this issue.

To make matters worse, there is a bug in 1.8 versions where if a block
is set in an empty chunk, that block is set in _all_ instances of the
empty chunk throughout the map.  With WDL this is worse, since the empty
chunk with garbage data would then be saved in incorrect locations, but
this effect can be observed in vanilla sometimes.