Fix archive name
Previously: "mod-1.11.2-baseedit-...".  Now, it's actually "mod-worlddownloader-baseedit" as it should be.  This happened because the FG plugin was applied before the base name was set.  (The java plugin needs to be applied before the base name can be set too, hence the weird order)

Additionally, the group was updated to be "wdl" in all cases, for convenience if I ever put the mod on maven (all versions will be in the same place).