Add gradle buildscripts for 1.12
This is, as always, somewhat incorrect.  1.12 requires a new version of fernflower (now renamed forgeflower).  While I can directly update it, if I update forgegradle directly, non-1.12 versions break (the new fernflower fixes some decompile errors, which cause patches to fail - in particular, it stops adding some useless imports, and that makes patches not line up).

But, for the moment, everything will decompile.  Slowly, but it does.  So this is a start.