Fix nonexistant class VertexBuffer
It was renamed to BufferBuilder in 1.12, a rename which I agree with (it
gets rid of a duplicate class name), so I'm backporting that rename.  I
haven't actually tested this change (it's a 2-hour rebuild, which I'd
rather merge with the mapping changes which will come later), but
everything should work.