Fix reobf, update mappings
Due to Pokechu22/ForgeGradle-extensions#2 (which I did not encounter with EquineEntity), I can't remap classes that I plan to call methods on.  So, I use LocalUtils instead.  This was a bit of a pain to diagnose, and I'll want to fix Pokechu22/ForgeGradle-extensions#2 soon, but in the meantime, I CAN RELEASE!  Soon, at least.