Completely update begradle; update mappings
The begradle update fixes #64 by referencing only the jar file for begradle, which causes gradle to not attempt to resolve dependencies (including the specific version of forgegradle, which doesn't work with all versions).  This does mean that other dependencies need to manually be specified, but that isn't too hard to do.