Disable source jars better
This removes the task as a prerequisite, rather than just disabling the task itself.  That in turn keeps the dependencies of the task from being run - making builds faster.

As a side bonus (not really useful, but worth noting), the sourceJar task can now be manually run, if anyone cares.  The issues with it still exist, but that doesn't matter.