Fix importing block entities
In 1.11, block entities transitioned from IDs in the form of `Chest`, `Dropper`, `Trap` to namespaced IDs such as `minecraft:chest`, `minecraft:dropper`, and `minecraft:dispenser`.  However, WDL was still using the old IDs.  This meant that block entities would never be imported in versions above 1.11.

This should be a fix for #60.