Fix displaying backup progress
Primarily, this fixes a major mistake I made in implementing IBackupProgressMonitor - setNumberOfFiles set the progress instead of the count.  But also, the percentage display for minor task progress was assigned to the major task progress, causing duplicate text and some text to not display.

This also removes the synchronized keyword in several places - volatile is better for this purpose as far as I understand.