Attempt to detect+fix the malformed block entity importing
My current guess is that saveChunk is somehow being called twice, which
is causing MC-119971-esque loading of the old version of the chunk.
I've added a check to see if the chunk is already in chunksToSave for
importing, which should help (but wouldn't be a complete fix, due to the
same principles as MC-119971 itself).  I've also added some more debug
logging to attempt to detect duplicate calls to saveChunk.