Make testCompile depend upon forgeGradleMcDeps
This currently does nothing, but a later change will add a dependency that depends on LauncherWrapper (LegacyLauncher), which in turn depends on some of MC's dependencies... but at an older version.  This causes weird issues where gradle adds both versions to the eclipse configuration, in turn breaking everything; including the MC deps avoids this issue.