Attempt to implement the inventorybasic stuff in the litemod
This messes up everything.  LWTS is supposed to be used so that I can use the unit tests, but unfortunately it doesn't work yet (weird issues abound).  And since LWTS is only used for the litemod, this breaks the base edit too.  But only for the unit tests, at least.

I'm commiting this now so that I can have a stable ground for some other work.