Add a fallback so that tests can still work when mixins aren't applied
The main part is to have code to check against the vanilla name(s) which is used if INetworkNameable does not work.  In addition I use JUnit's Assume function to skip the tests that would _require_ the mixins to behave accurately.  As such I can still run most of the custom name tests, just skipping the ones that truely cannot work.