Redo the way I18n works in tests
The old method worked by delegating to the legacy I18n class, which accessed MC's language files (assets/minecraft/lang) but not WDL's (assets/wdl/lang).  This meant that WDL text couldn't be used in tests, which is silly.

The new method is to properly set up the language manager, using the DefaultResourcePack (which is already modified to access WDL's namespace).  There actually is no noticable performance cost to doing this.

However, it doesn't work on liteloader, which I neglected to test until just now.  I will look for a solution...