Move construction of SPacketCustomPayload to VersionedProperties
In 1.13, they take a ResourceLocation, so this is needed.

Also, the UTF-8 error can never happen; UTF-8 is guaranteed to be
present (after all, parts of the class file are encoded in UTF-8...).
Using StandardCharsets (which I must not have found) lets us avoid the
need to deal with an UnsupportedEncodingException.