Update plugin message handler for 1.13
1.13 mandates that channel names be namespaced, e.g. minecraft:book_open
instead of MC|BOpen.  This means that WDL's channels must be renamed.
Additional complications come from the fact that we cannot easily tell
if the server's using the new names or the old names without listening
for REGISTER (now minecraft:register), so now that is done as well.  One
pain point is that there is a chance that the server hasn't sent the
register packet by the time we check, so there is functionality to defer
sending the init packet until a register packet arrives.  However, this
functionality has not been tested in practice, since on my machine
register always does arrive in time.