Add missing 1.13.1 realms bridge methods
This fixes crashes due to NoSuchMethodErrors on realms screens, though ideally the methods in question won't be called either.

hasWidget is the most notable method of this set, as it's the only one that actually is called (I think).  The others are to fix some thread safety issues that cause JVM crashes (I think) -- which do matter, but are less important.