Fix broken [WorldDL] text tooltip
Fixes #95

Note that this text has \n inside of it only so that it can have decent
word-wrapping; in 1.12 the text is wrapped automatically so it's less
important.  However, I still need this manual handling for earlier
versions.  I considered wrapping the text myself in older versions, but
that introduces a dependency on the font rendering code which is
dangerous to access from other threads (which WDLMessages must support).

This is also tied to Pokechu22/ForgeGradle-extensions@7f57329 and
Pokechu22/ForgeGradle-extensions@f8ec7d2.