Fix #99 - issues with multiple things reading the same buffers
The issue is that both liteloader and WDL read from the same buffer, which created issues if liteloader read the whole buffer and then WDL saw that it was (now) empty.  In addition, this probably caused issues with other mods, though I haven't fully confirmed this.

37dc67cd7577182703633992922efb8a8b1308b8 was when the readable check was added, to fix a crash when interacting with villagers.  I later changed from using array() to using 3fae16151960351f9bdb2d265419dd2ac285018b to get the buffer contents.  The villager check actually prevented a crash along the lines of this:

```
io.netty.util.IllegalReferenceCountException: refCnt: 0
	at io.netty.buffer.AbstractByteBuf.ensureAccessible(AbstractByteBuf.java:1408)
	at io.netty.buffer.AbstractByteBuf.checkReadableBytes0(AbstractByteBuf.java:1394)
	at io.netty.buffer.AbstractByteBuf.checkReadableBytes(AbstractByteBuf.java:1383)
	at io.netty.buffer.AbstractByteBuf.readBytes(AbstractByteBuf.java:850)
	at io.netty.buffer.AbstractByteBuf.readBytes(AbstractByteBuf.java:858)
```

Note that this is _not_ an issue with reading beyond the end (which is what I thought when I added that).  This is rather an issue when the reference count hits 0... which happens when some packets are released (this happens inconsistently with vanilla packets in older versions; see https://bugs.mojang.com/browse/MC-121884).  I've changed this to actually checking the reference count, which is the crux of the issue.  (We can't recover from the refCount hitting 0 as the packet is freed at that point, but we _can_ reset the reader index).

Setting the reader index to 0 for the purposes of re-reading the contents is always the correct location as plugin message's buffer is always created by `new PacketBuffer(buf.readBytes(i))` and the documentation for readBytes states that "The returned buffer's readerIndex and writerIndex are 0 and length respectively."  Older versions might behave slightly differently, but I think this guarantee will always hold.

To add on to things, though, Forge also intercepts REGISTER packets, and _it_ never sends it on to the NetHandlerPlayClient (nor calls processPacket even).  To resolve this, I inject a netty handler before forge's' handler which relays the data to WDL.  However, this would have the complication of sending data to WDL twice if forge was not installed, so I check whether forge's handler is present on the first packet and remove my handler if it is not (we cannot merely check if the handler is present at the time when we add it, due to version differences - see https://git.io/v7Lvp).  We do check if forge's handler is already known to be present and add before it if that is the case, but if it isn't then we add before the vanilla handler; forge also adds its own before the vanilla handler which would be after our handler in that case.

Unfortunately this does produce a lot of duplicated code, but that's how the litemod-specific code is anyways so I'm not going to worry about it much.

Note that I previously used a pseudo mixin into NetworkDispatcher to add code before https://git.io/fxaRK to relay to WDL.  This works for 1.12, but injecting into FML's code feels dirty, and also we don't have access to pseudo mixins in 1.10.2 and earlier so in the long run that doesn't actually work.

Tested cases:
* Vanilla REGISTER
* WDL's handshake
* The trade list packet on villagers (which isn't a plugin message on 1.13, but this wouldn't have applied to 1.13 as it currently stands anyways)

Tested versions:
* 1.9 base-edit
* 1.9 litemod
* 1.9 litemod, with forge also installed
* 1.12.2 base-edit
* 1.12.2 litemod
* 1.12.2 litemod, with forge also installed