Fix disconnects not being detected in some cases (#100)
The mod had code for detecting these cases a long time ago, but it was
lost between 1.10 and 1.11.  Moreover, it was lost again when I
backported v4 to those versions, so only the litemod had it.

One important thing to note is that we don't need to handle the
disconnect packet, as that defers to onDisconnect instead.

This should fix the cases in #100 where disconnects weren't properly
detected.  I _still_ do not feel 100% comfortable with these unusual
disconnect situations, and recommend always using the stop download
button to be sure.  More unit tests will hopefully help resolve these
concerns.