Fix NPE in deferred init (fixes #104)
The same kind of weirdness with early messages from bungee before that caused Minecraft.getConnection() to be null (7bfb3f96) could happen here when there was already a deferred init.  The remaining uses of getConnection should be fine (sending requests and the external use of sendInitPacket should both be ingame, and onWorldLoad by definition is ingame).  Hopefully, at least...