Capture and display external backup command output
This is a bit janky, but it works.  Only tested on windows, not linux
(7-zip does give different output there, so that may cause issues).

For it to behave usefully, -bsp1 must be produced to 7-zip.  Otherwise
it doesn't have anything I can parse, and more importantly output isn't
flushed until the program finishes.