Change the gamerules GUI to modify the actual rules for the world
This means loading the existing rules and modifying them, instead of
overriding the defaults on save.  i.e. it's possible now to use
/gamerule in singleplayer and that will also work, although the UI is
preferable since you can change things before the world loads.  See #111
for more info.

I haven't tested the code yet.  I do know that the behavior isn't
completely great in terms of when they're actually written -- if the
values are changed in the GUI but the world isn't saved afterwards, the
changes are ignored (unlike other GUIs).  Fixing that is complicated due
to possibly needing to create a new level.dat from scratch, which I'm
just not sure how to do without using weird defaults.  That's something
for the eventual multiworld cleanup, I guess...