Fix conflict with MoreCommands
MoreCommands does some interesting stuff including modifying NHPC by
subclassing it, but the way it does that leads to both a regular and a
patched NHPC instance being made, which caused my code to be called
twice.  While this weird behavior probably breaks other mods too, I can
at least handle it by checking if I've injected a packet handler
already.

Injecting twice threw an exception, but due to MC-133130 this exception
wasn't logged properly and a later issue with forge trying to inject
_its_ packet handler was instead logged.  Yay for debugability :/

Soccerguy is currently working on morecommands, and I don't think there
actually is a finalized public build for testing yet (but I've been
helping him with it, so I've got the needed changes to test).