Remove duplicated code in multiworld callbacks
This isn't the best way of approaching it at all; it'd be better to have
some kind of prompt chain in a single class that asks for things as
needed (enable multiworld, which world, overwrite changes, etc).  I do
think the attempting to do everything a second time approach is cleaner
than copying each inner step (especially since canceling the second step
doesn't undo the first step), though the hacky way I do that in initGui
is suboptimal.