Fix a few issues with the saved chunks GUI (#115)
createOrLoadRegionFile takes a chunk coordinate, not a region
coordinate.  I can probably optimize this a bit and just depend on the
region cache to load things properly for me instead of manually getting
region files, though repeatedly calling something like that seems
painful (and I still don't want to create empty files!)

Also, if the timestamp is 0, the chunk shouldn't be rendered since it
hasn't been saved (or someone magically time-traveled back to 1970 to
save it... which seems unlikely to me).

There's still a lot of additional work to be done on this.