Use HSL algorithm to make colors better
Granted, in this case, S and L are constant, so it's just hue and
relatively simple.  The key is to not try to balance r+g, but just let
one increase and then later the other decrease -- we don't even need the
giant modulo logic thing because we only care about the orange and
chartreuse regions.