Remove obviously irrelevant versioned files from 1.14
If things get merged together, we don't need 2 copies of these since
they're ignored.  (Possibly the 1.14 and 1.13 versions of some things
will be merged too, but that hasn't happened yet)

Also, I just realized that even if things get merged, this whole mess
will be captured in git history, so it'll increase the size of the git
repo.  Nothing can be done about that, really, but it's still going to
be a pain.

Errors: 2229.