Fix data files not being on the classpath
This is a bit of an odd fix.  However, it does mean that tests stop all
erroring due to Bootstrap not working from the missing language data.
With proper WDL language files (which might not be present in the IDE,
but will be when running with gradle or if the json converted file is
put in the resources folder - an awkward and unrelated thing), the test
statistics are:

1.14.0 Runs 225/225 (17 skipped) Errors: 147 Failures: 0
1.14.1 Runs 225/225 (17 skipped) Errors: 148 Failures: 0
1.14.2 Runs 225/225 (17 skipped) Errors: 148 Failures: 0

Note that for some reason 1.13.2 has WDLHooksTest erroring, which
persists to here.  I'll check that.