Work around issues with WDLHooksTest
The problem is that the minecraft instance isn't updated before
WDLHooks' static initializer is called if WDLHooksTest runs before
another thing that uses AbstractWorldBehaviorTest (reproducible on any
version if it's the only test that runs, and because test order is
nondeterministic it occurs in 1.13 and 1.14 for me).  My workaround is
to just move WDL's usage of the Minecraft instance to inside a try
block, which ignores the issue (the actual use of WDL is skipped later
on; it's only the static initializer that references it).

WDLHooksTest is still broken in 1.14, but not because of this at least.