Update patches for 1.09.2 and 1.09.4
Thus ends the giant updating process.  Wasn't as bad as I thought it
was... but of course if I want to reconcile class names (which I will
need to do eventually) it'll be like this but 10 or 15 times as bad due
to the increased time spent for decompiling.  So... yeah, that'll be
fun.