Switch 1.13.x to ForgeGradle 3
A big step in the process of solving #116.  Unfortunately this still
isn't complete, as the extra SRGs and CSVs don't work yet.  This means
there's a tone of errors for the mod itself.

Fortunately, the different 1.13 versions recompile the base game
cleanly.  It's only 1.14.0 that has issues with recompiling.  Since
1.14.0 is less likely to be used, I'm not too worried about it (and once
I fix the patching infrastructure, perhaps I can get that solved too!)