Update 1.14.4 files to be for 1.14.4
The one thing that's a bit tricky here is the library list in 1.14.4/build.gradle, which hopefully will not be needed in the future.  I generated it by running the following:

    curl https://launchermeta.mojang.com/v1/packages/132979c36455cc1e17e5f9cc767b4e13c6947033/1.14.4.json | python3 -m json.tool | grep name | clip

... and then manually editing the result to look correct (and removing the duplicate and/or macos only libraries).

I also ran `gradlew regenEclipse` twice, once to decompile and a second time to recompile.  That also shouldn't be needed in the future (setupDecompWorkspace should do the job instead), but it currently is.

Note that code doesn't completely compile as-is, either.  This is just setting things up.

('Documented' for #130)