Fix 1.14.4 compile errors
In this case, the errors were caused by a (mandatory due to technical reasons on the MCP side) package change, and except for one caused by a class usage in a javadoc comment, they were all in test code.

I added more per-version code to deal with this, including ignoring it in build.gradle.  I then ran `./gradlew :1.14.{1,2,3,4}:regenEclipse` to update my IDE configuration (and manually set up 1.14.0 because of how fickle it is currently).  The actual per-version code is a bit hacky, but not much hackier than it already was.

Running the unit tests shows everything as passing except for SanityCheckTest and WDLHooksTest, which both check base code changes.  Since the base patches haven't been updated yet, this is to be expected.

('Documentation' for #130)