Get things to boot
Apparently the list of mixins is a command-line arg now?  No idea how that'll work in production.

In addition to that, note that ClientWorld.removeEntityFromWorld no longer returns an entity, so MixinWorldClient had to be tweaked.  I'm only noting that because I haven't added that code yet (because the mixin code is super messed up at the moment and will need to be tidied first).