Resolve errors due to tags not being constructible
The new unnamed methods are valueOf or similar.  Tags now have a cache
for values (similar to Integer/Byte/other boxed classes), so the method
needs to exist to allow use of that.  I backported the same versioned
code to 1.13 for convenience (and to move the nbtString method to a
better place).

Down to 133 errors.