Partially fix GlStateManager changes
Almost all of these are just functions missing mappings, and can be
undone later.  (There are a few changes to the parameters to tex() being
floats instead of doubles, but those can be backported)

Note that the functions are deprecated.  I'm not sure what the intended
replacement is, so I won't touch it yet.

95 errors remain.