Remove duplicate NBT writing in WDLChunkLoader_1_15
This also set the wrong status.  In previous versions, that would cause
a new world to generate over existing chunks (if generation was
enabled), but that doesn't seem to be happening here (I'm not sure why).
It *did* increase loading times of the downloaded world, though, since
it went through a bunch of generation phases (seen by lots more colors
on the loading progress indicator).