Make WDL.minecraft not static
To explain these changes: 1.15.2 creates a CrashReport instance on
startup.  That in turn loads WDL, but it happens before the Minecraft
instance is set... which causes massive issues.  So I need to refactor
the way WDL is loaded to resolve that.

(Similar refactoring is probably a good idea to be in line with the
order Forge wants)