Refactor hooks instance handling
This actually fixes 1.15.2, but I want to do some more refactoring
first.  (Plus, this needs to be backported).

WDLEvents creates a WDLHooks listener, and both of those are passed a
WDL instance, instead of both of them statically going backwards and
then finding the WDL instance.  The default hooks handler will bootstrap
WDL, creating the appropriate events and hooks listeners.  Other loaders
can bootstrap at different times (as is done with forge), but the
default handler is used as a fallback for the base edit.