Clear last block pos if no block is targeted on window open
If a chest was opened and saved, but then another chest was saved while
not looking at any block, that would cause the last chest's position to
be used.  (Prior to 7df1d1efa3 it would cause a crash.)

Fixes #137.  Fixes #150.  The state handling with those two fields is
still somewhat messy, but this is still a correct implementation and
further improvements will build on it (instead of removing it).