Add field_150278_b workaround
The 20200213 mappings added back comments for some old things, and this one was mangled (with some unmatched quotes) in a way that caused FG2 to choke (and rename next to no fields).  I've fixed the comment, but that won't propagate until tomorrow.  In the mean time, I get to recompile everything again...