Use META-INF MixinConfigs for the forge mod
This requires MixinBootstrap to be installed (for now), but that's a lot
easier to install than the previous jank method of manually adding mixin
to the libraries list in a version JSON and adding extra arguments there
as well.

Also included is a message that informs the user to install
MixinBootstrap if sanity checks fail, so that they can fix it instead of
being left wondering why the pause menu hasn't changed.