Rework test bootstrap
Instead of having both NoMixin and Mixin versions of each versioned
file, there now is a versioned file, and a smaller file that controls
whether mixin should be used or not.

The TestBootstrap files are based on the MixinTest files.  There may be
further adjustments that can be done to them...