Fix Villager test with Forge
Forge makes it so that sneaking prevents opening the villager UI.  This
isn't a new change, but since these builds use forge directly (as
opposed to testing with liteloader only with forge possibly existing at
runtime), we need to not sneak to avoid running into that.

Only the entity utils/managers tests are broken now.