Fix #176
enqueue() always schedules a task to be run later, even if on the execution thread.  And, after disconnecting, MC clears all of the queued tasks, so the task would never actually be ran.  In the case of this issue, it was called from the execution thread, so we need this immediate call, which execute() provides.

In versions prior to 1.13, addScheduledTask (func_152343_a) is used.  For convenience I have that renamed to match what I used in current versions, but it was a different method that actually did handle the immediate execution case.

The off-thread case is still slightly dubious, but I think the sleep call will handle it (surely nothing can go wrong by assuming the scheduled task handler will run within 2 seconds...)