Only print one message for startup sanity check
Displaying all of the failed mixin messages may overwhelm people, and
they're largely redundant.  I've also tweaked the message to say "mixin
changes" instead of "base changes", and made the mixin sanity check be
the first one (as if multiple things are broken, we want the mixin
message to be the one to show in that case).