Fix NPE on save screen
It's inconsistent, but chunkLoader could be null in some cases, in which
case the subtitle would trigger an NPE.  This only could happen in 1.13+
since only in 1.14+ does chunkLoader get set to null.  I think this was
the same crash as in #176, though I don't think it was directly related
to getting disconnected.