Fix the random ConcurrentModificationException crash in 1.16
For some reason, closing the chunk loader without first flushing it can
cause CMEs in 1.16.  We already wanted to flush, but there wasn't an
obvious way in 1.14; func_227079_i_ (to be renamed at some point)
appears to handle flushing in 1.15+ though.

It's hard to prove a negative, though, so this *might* not actually fix
the CMEs.  I haven't ran into any yet, though.