Partially fix superflat settings
This seems like the right idea (RegistryLookupCodec wants
WorldSettingsImport or else it logs the warning "Not a registry ops"
(probably a mapping oddity)), and with this change settings can be saved
and loaded.  However, the biome part does not work, with the biome
showing up as "null" in the preset and "Received invalid biome id: -1"
being spammed in the log for generated chunks.  Biomes are still saved
as their contents instead of the key for some reason.