Fix players spawning in the wrong dimension
The issue was that Entity no longer stores the dimension; instead it
happens in ServerPlayerEntity, but WDL by necessity works on a
ClientPlayerEntity.  Much simpler than all the dynamic registries stuff
(which still should be cleaned up at some point, because I'm guessing
neat stuff could be done with it, but I don't think it needs to be
messed with just now).

All that's left, I guess, is fixing buffet.